Add price information for all tokens
Abstract getPriceInformation out to it's own file, change file structure to make more sense, add loader to table.
