+++ src/components/ERC20Balance/balanceMethods/multicall.ts
@@ -2,7 +2,7 @@ import { Fragment, Interface, JsonFragment } from "@ethersproject/abi";
 import { Contract } from "@ethersproject/contracts";
 import { providers } from "ethers";
 import { BigNumber, Bytes } from "ethers";
-import { getChainDetails } from "../helpers/getChainDetails";
+import { getChainDetails } from "../../../helpers/getChainDetails";
 import MULTICALL_ABI from "./multicallAbi.json";
 
 const MAX_MULTICALL_SIZE = 100;
