+++ src/components/ERC20Transfers/ERC20Transfers.jsx
@@ -3,7 +3,6 @@ import { useMoralis } from "react-moralis";
 import { getEllipsisTxt } from "../../helpers/formatters";
 import { getExplorer } from "../../helpers/networks";
 import "antd/dist/antd.css";
-// import { Table } from "antd";
 import { useERC20Transfers } from "../../hooks/useERC20Transfers";
 import Table from "../reusable/Table";
 
@@ -55,16 +54,6 @@ function ERC20Transfers() {
 
   return (
     <div>
-      {/* <Skeleton loading={!ERC20Transfers}>
-        <Table
-          dataSource={ERC20Transfers}
-          columns={columns}
-          rowKey={(record) => {
-            key++;
-            return `${record.transaction_hash}-${key}`;
-          }}
-        />
-      </Skeleton> */}
       <div
         style={{
           margin: "0 3.125rem 3.125rem 3.125rem",
@@ -72,6 +61,7 @@ function ERC20Transfers() {
         }}
       >
         <Table
+          loading={!ERC20Transfers}
           tableData={ERC20Transfers}
           columns={columns}
           tableTitle={"Transactions History"}
