+++ src/components/NativeBalance.jsx
@@ -5,7 +5,12 @@ function NativeBalance(props) {
   const { data: balance, isLoading } = useNativeBalance(props);
   const { account, isAuthenticated } = useMoralis();
   if (!account || !isAuthenticated) return null;
-  if (isLoading) return <Spin style={{ color: "#3e389f" }}></Spin>;
+  if (isLoading)
+    return (
+      <div style={{ textAlign: "center", marginTop: "1.25rem" }}>
+        <Spin style={{ color: "#3e389f" }}></Spin>
+      </div>
+    );
   return (
     <div
       style={{
